Module Engine.Step

type t

A unique ID representing the current iteration. This is used by the cache to warn about attempts to set the same output to two different values at the same time.

val equal : t -> t -> bool

val now : unit -> t