Module Current.Primitive

An OCurrent pipeline is made up of primitive operations. A primitive is roughly the content of a single box in the diagram.

Warning: Primitive is the low-level API. You will almost always want to use Current_cache (for processing or publishing jobs) or Monitor (for inputs) instead.

type 'a t = ('a Current_term.Output.t * Metadata.t option) Current_incr.t

val const : 'a -> 'a t

const x is a primitive that always evaluates to x and never needs to be updated.

val map_result : ( 'a Current_term.Output.t -> 'b Current_term.Output.t ) -> 'a t -> 'b t

map_result fn t transforms the result of t with fn. The metadata remains the same. If fn raises an exception, this is converted to Error.