Module Current.Step

type t

type id

val id : t -> id

id t is a unique value for this evaluation step. This can be useful to detect if e.g. the same output has been set to two different values in one step.

val config : t -> Config.t