Class type Current.actions

method pp : Stdlib.Format.formatter -> unit

Format a message for the user explaining what is being waited on.

method rebuild : (unit -> job_id) option

A function to call if the user explicitly requests the operation be done again, or None if it is not something that can be repeated. Returns the new job ID.