Module Current_cache

Cache build results in memory and on disk. A cache maps keys to values. Looking up a key that isn't known starts a new build to create it.

module S : sig ... end

module Schedule : sig ... end

Configuration settings controlling when to rebuild.

module Make (B : S.BUILDER) : sig ... end

Perform builds with caching.

module Output (P : S.PUBLISHER) : sig ... end

Publish outputs with caching.

module Generic (Op : S.GENERIC) : sig ... end

The most general API. Make and Output just specialise this for particular uses.

module Db : sig ... end

Low-level database access (for the web UI and plugins).