Module Current_cache.Generic

The most general API. Make and Output just specialise this for particular uses.

Parameters

Op : S.GENERIC

Signature

val run : ?⁠schedule:Schedule.t -> Op.t -> Op.Key.t -> Op.Value.t -> Op.Outcome.t Current.Primitive.t

run t k v is a term for the result of processing (k, v).

val reset : db:bool -> unit

reset ~db clears the cache. Useful for unit-tests.

parameter db

Also clear from the on-disk database table.