Parameter Make.B

include S.OPERATION

type t

Extra context or configuration information (e.g. credentials to access a remote service).

val id : string

A short and unique name for this operation, used as a filename component and database key. e.g. "docker-build"

module Key : S.WITH_DIGEST

The id, Key.t pair uniquely identifies the operation.

module Value : S.WITH_MARSHAL

The result of a build.

val build : t -> Current.Job.t -> Key.t -> Value.t Current.or_error Lwt.t

build t j k builds k. Call Job.start j once any required resources have been acquired. Log messages can be written to j.

val pp : Key.t Fmt.t

Describe the operation. e.g. "docker build $src".

val auto_cancel : bool

true if an operation should be cancelled if it is no longer needed, or false to cancel only when the user explicitly requests it.