Module type S.PUBLISHER

A publisher sets a key to a value.

include OPERATION

type t

Extra context or configuration information (e.g. credentials to access a remote service).

val id : string

A short and unique name for this operation, used as a filename component and database key. e.g. "docker-build"

module Key : WITH_DIGEST

The id, Key.t pair uniquely identifies the operation.

module Value : WITH_DIGEST

The value to publish.

module Outcome : WITH_MARSHAL

Extra information about the result, if any. Usually this is just Current.Unit.

val publish : t -> Current.Job.t -> Key.t -> Value.t -> Outcome.t Current.or_error Lwt.t

publish t j k v sets output k to value v. Call Job.start j once any required resources have been acquired. Log messages can be written to j.

val pp : (Key.t * Value.t) Fmt.t

Describe the operation. e.g. "docker push $tag".

val auto_cancel : bool

This affects what happens if we're in the process of outputting a value and then we decide to output a different value instead. If true, the old operation will be cancelled immediately. If false, the old operation will run to completion first.