Module BUILDER.Value

The result of a build.

type t

val marshal : t -> string

Convert t to a form suitable for storage in the database.

val unmarshal : string -> t

Restore a t from a string previously produced by marshal. Raise an exception if the value cannot be read (e.g. the format has changed).