Module type S.ANALYSIS

type 'a term

See TERM.

type metadata

Extra data provided by primitives but that isn't part of the value that is passed on to other nodes. For example, a job ID.

val metadata : 'a term -> metadata option term

metadata t is the metadata of t, if any. Raises an exception if t is not a primitive (or a map of one).

val pp : _ term Fmt.t

pp formats a t as a simple string.

val pp_dot : env:(string * string) list -> url:(metadata link -> string option) -> _ term Fmt.t

pp_dot ~env ~url formats a t as a graphviz dot graph.

parameter env

A list of key-value pairs from the URL to control rendering.

parameter url

Generates URLs for links.

val stats : _ term -> stats

stats t count how many stages are in each state.