Module type Current_term__S.INPUT

type 'a t

An input that was used while evaluating a term. If the input changes, the term should be re-evaluated.

type job_id

type env

A context which the caller can associate with an execution.

val get : env -> 'a t -> 'a Current_term.Output.t * job_id option