Module Current.Config

type t

val v : ?⁠auto_release:Duration.t -> ?⁠confirm:Level.t -> unit -> t

A new configuration.

parameter auto_release

Remove confirmation requirement this period (unless changed manually first).

parameter confirm

Confirm before performing operations at or above this level.

val set_confirm : t -> Level.t option -> unit

Change the confirm setting. Existing jobs waiting for confirmation will now start if permitted by the new configuration.

val get_confirm : t -> Level.t option

val cmdliner : t Cmdliner.Term.t