Module Current_cache.Db

type entry = {

job_id : string;

build : int64;

value : string;

outcome : string Current.or_error;

ready : float;

running : float option;

finished : float;

rebuild : bool;

}

val init : unit -> unit

Ensure that the database tables have been created. This is useful if you need to refer to them in your own SQL.

val query : ?⁠op:string -> ?⁠ok:bool -> ?⁠rebuild:bool -> unit -> entry list

Search the database for matching records.

parameter op

: if present, restrict to results from the named builder or publisher

parameter ok

: if present, restrict results to passing (ok=true) or failing (ok=false) results.

parameter rebuild

: if present, restrict results to ones where the rebuild flag matches this.