Module Current_cache.Db

Low-level database access (for the web UI and plugins).

type entry = {

job_id : string;

build : int64;

value : string;

outcome : string Current.or_error;

ready : float;

running : float option;

finished : float;

rebuild : bool;

}

val init : unit -> unit

Ensure that the database tables have been created. This is useful if you need to refer to them in your own SQL.

val query : ?op:string -> ?ok:bool -> ?rebuild:bool -> ?job_prefix:string -> unit -> entry list

Search the database for matching records.

parameter op
: if present, restrict to results from the named builder or publisher

parameter ok
: if present, restrict results to passing (ok=true) or failing (ok=false) results.

parameter rebuild
: if present, restrict results to ones where the rebuild flag matches this.

parameter job_prefix
: if present, restrict results to ones where the job ID starts with this string.

val ops : unit -> string list

ops () is the list of operation types that can be passed to query.

val history : limit:int -> job_id:string -> string option * entry list

history ~limit ~job_id returns the in-progress build (if any), and the limit most recent builds with the same key as job_id.