Module PUBLISHER.Key

The id, Key.t pair uniquely identifies the operation.

type t

val digest : t -> string

digest t is a unique string for t that can be used as a primary key in a database. Two ts are considered equal if they have the same digest.